refactor(map): split view-builder markers suite under the 4-file cap
Fold visibility/presence into region_cells and move town/civilian marker
coverage into its own suite so AC5 of #4022 is met without exceeding
repo.map_test_view_builder_file_count.
