+++ SPEC/program/repo-lint.md
@@ -122,7 +122,7 @@
 | `tool/check_orders_scenario_table_runner.dart` | Prefer scenario-table runners for new long imperative `test()` bodies under `packages/colonizethis_orders/test/**`; after wave-3 slice 13, baseline allow-all is off and remaining imperative suites are listed in `ordersPreferScenarioTablesAllowlist` / `REFACTOR_TRACE.md` (Refs #3949); rule `repo.orders_test_prefer_scenario_tables` |
 | `tool/check_combat_test_no_duplicate_fixtures.dart` | Forbid re-introducing a local copy of the shared game/world test fixtures anywhere under `packages/colonizethis_combat/test/**`: any file named `test_fixtures.dart` (the deleted byte-identical-modulo-doc-comment subset of `package:colonizethis_test/game_test_fixtures.dart`) or any test file that redeclares a top-level `TestFixtures` class (optionally `abstract`/`final`/`base`/`sealed`/`interface`/`mixin`) is a violation. Combat tests must import the canonical `package:colonizethis_test/game_test_fixtures.dart` factories (single source of truth) rather than duplicating them; references to the shared `TestFixtures` symbol are unaffected, and classes whose names merely contain `TestFixtures` (e.g. `MyTestFixturesBuilder`) are not flagged. Complements the testing-rule shared-fixtures policy (Refs #3716); rule `repo.combat_test_no_duplicate_fixtures` |
 | `tool/check_map_test_no_duplicate_view_fixtures.dart` | Forbid re-inlining the dual-region `buildInitGameMapViewData` test scaffolding under `packages/colonizethis_map/test/init_game_map_view_builder_*`: a file in scope is a violation when it renders a view (`buildInitGameMapViewData(`) **and** still constructs the dual-region scaffolding inline — an inline `Game(` (`\bGame\(`, so `minimalGame(` is not matched) or `MapTopology(` (`\bMapTopology\(`, so `regionTopology(` / `singleProvinceAndSeaTopology(` are not matched). View-builder tests MUST build their `owMap`/`nwMap`/`owTopology`/`nwTopology`/`Game` setup through the shared `test/support/init_game_map_view_fixtures.dart` builders (`minimalGame`, `regionTopology`, `singleProvinceAndSeaTopology`, `mapTileGrid`); full-line comments are stripped so prose mentioning the constructors is not flagged, and the shared support file plus other map suites (e.g. `tile_map_visualization_test_fixtures.dart`, generator tests with bespoke grids) are out of scope (Refs #3746); rule `repo.map_test_no_duplicate_view_fixtures` |
-| `tool/check_map_test_view_builder_file_count.dart` | Hard cap of **4** files matching `packages/colonizethis_map/test/init_game_map_view_builder_*_test.dart` after the #3846 view-builder consolidation; no documentation escape hatch (Refs #3846); rule `repo.map_test_view_builder_file_count` |
+| `tool/check_map_test_view_builder_file_count.dart` | Hard cap of **4** files matching `packages/colonizethis_map/test/init_game_map_view_builder_*_test.dart` after the #3846 view-builder consolidation; no documentation escape hatch (Refs #3846). After #4022 AC5, the former markers mega-suite is split (`*_markers_test.dart` vs `*_town_civilian_markers_test.dart`) by folding visibility into `*_region_cells_test.dart` so the file count stays ≤4; rule `repo.map_test_view_builder_file_count` |
 | `tool/check_map_test_shared_topology_fixtures.dart` | Forbid inline `\bTopologyNode\(` in render/visualizer tests (`test/tile_map_visualization_*`, `test/game_world_state_map_visualizer_test.dart`, `test/multi_region_map_rendering_test.dart`) unless allowlisted: hardcoded exempt files (e.g. `tile_map_topology_validation_test.dart`), `*_validation*_test.dart` basename glob, or a per-site `// map-topology-fixture-exempt: <reason>` marker on the line immediately preceding the inline constructor; `test/support/**` and `test/*_fixtures.dart` are out of scope (Refs #3846); rule `repo.map_test_shared_topology_fixtures` |
 | `tool/check_map_test_run_generation_harness.dart` | Forbid inline `\bTileMapGenerator\s*\(` under `packages/colonizethis_map/test/tile_map_generator_*_test.dart`, `tile_map_generation_*`, `tile_map_pass4_*`, `tile_map_forest_*`, and `tile_map_terrain_*` suites unless the preceding line carries `// map-generation-harness-exempt: <reason>` (constructor/DI probes). Integration generates must use `runTileMapGeneration` from `test/support/tile_map_gen_fixtures.dart`; that support file is out of scope (Refs #4022); rule `repo.map_test_run_generation_harness` |
 | `tool/check_map_test_minimal_game_shared.dart` | Forbid inline `\bGame\(` in the scoped colour/capital/region-input/format/port suites (`faction_ownership_color_test.dart`, `tile_map_capital_markers_test.dart`, `region_map_view_inputs_test.dart`, `map_format_util_test.dart`, `port_icon_placement_test.dart`); those suites must use `minimalGame` from `test/support/init_game_map_view_fixtures.dart`. Full-line comments are stripped; `minimalGame(` is not matched (Refs #4022); rule `repo.map_test_minimal_game_shared` |
@@ -522,6 +522,7 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given the generation-layer files under `packages/colonizethis_map/lib/src/gen/**`, view-layer files under `lib/src/view/**`, and render-layer files under `lib/src/render/**`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_lib_file_size`, then every scanned `.dart` file has at most 500 non-comment lines and the run exits `0` (Refs #3588, #3846).
 - Given a generation-layer, view-layer, or render-layer `.dart` file with strictly more than 500 non-comment lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_lib_file_size`, then the run fails listing the offending path with its non-comment line count and exits `1` (Refs #3588, #3846).
 - Given more than 4 files matching `packages/colonizethis_map/test/init_game_map_view_builder_*_test.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_view_builder_file_count`, then the run fails listing each matching path and exits `1` (Refs #3846).
+- Given the map view-builder marker suites after #4022 AC5, when counting `init_game_map_view_builder_*_test.dart` files, then the count remains ≤4 and ownership/capital/port/warp marker coverage lives in `init_game_map_view_builder_markers_test.dart` while town/civilian marker coverage lives in `init_game_map_view_builder_town_civilian_markers_test.dart` (visibility/presence coverage is folded into `init_game_map_view_builder_region_cells_test.dart`) (Refs #4022).
 - Given `packages/colonizethis_map/test/tile_map_visualization_*`, `test/game_world_state_map_visualizer_test.dart`, or `test/multi_region_map_rendering_test.dart` with an unallowlisted inline `\bTopologyNode\(`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_shared_topology_fixtures`, then the run fails listing each offending path and exits `1` (Refs #3846).
 - Given an in-scope map generator suite (`tile_map_generator_*`, `tile_map_generation_*`, `tile_map_pass4_*`, `tile_map_forest_*`, `tile_map_terrain_*`) with an unallowlisted inline `\bTileMapGenerator\s*\(`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_run_generation_harness`, then the run fails listing each offending path and exits `1` (Refs #4022).
 - Given a scoped colour/capital/region-input/format/port map test with an inline `\bGame\(`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.map_test_minimal_game_shared`, then the run fails listing each offending path and exits `1` (Refs #4022).
